requirements.txt
openai
PyGithub
pydantic
pydantic-settings
python-dotenv